Fix some additional visualization tests
